backend/llm_engine/requirements.txt
openai==1.18.0
python-dotenv==1.0.1
grpcio==1.71.0
grpcio-tools==1.71.0
grpcio-health-checking==1.71.0
httpx~=0.27.0